pyproject.toml
[project]
name = "autocad-mcp"
version = "3.0.0"
description = "MCP server for AutoCAD LT automation and headless DXF generation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.1,<2.0",
    "ezdxf>=0.18",
    "matplotlib>=3.7",
    "Pillow>=10.0.0",
    "structlog>=24.0.0",
    "pywin32>=305; sys_platform == 'win32'",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autocad_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
